# console/openapi.yaml
openapi: 3.1.0
info:
  title: Trestle Operator API
  version: 0.2.0
  description: Sessions, telemetry, registry, and operator actions — in-process ControlSurface projection (/ops/v1).
servers:
  - url: http://127.0.0.1:18733
paths:
...
  /ops/v1/telemetry/chunk:
    post:
      operationId: read_telemetry_chunk
      summary: Bounded telemetry chunk (ControlSurface.fetch)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle, window]
              properties:
                handle:
                  type: string
                window:
                  type: object
      responses:
        "200":
          description: FetchSlice envelope or projection refusal
...
  /ops/v1/retention/{handle}:
    put:
      operationId: pin_retention
      summary: Pin session or artifact for retention (ControlSurface.pin)
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RequestOutcome envelope
    delete:
      operationId: unpin_retention
      summary: Remove retention pin (ControlSurface.unpin)
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RequestOutcome envelope
...